Modify AWS RDS

+++ Models/Store.cs
@@ -5,6 +5,6 @@
     public required int StoreId { get; set; }
     public required string Name { get; set; }
     public required string Image { get; set; }
-    public required int Distance { get; set; }
+    public required double Distance { get; set; }
   }
 }
